AppとRenderをwgpu_helperで提供
