+++ lib/wgpu_helper/Cargo.toml
@@ -4,6 +4,8 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-cgmath = "0.18.0"
-wgpu  = "22.1.0"
-winit = "0.30.5"
+anyhow   = "1.0.89"
+cgmath   = "0.18.0"
+pollster = "0.3.0"
+wgpu     = "22.1.0"
+winit    = "0.30.5"
