+++ lib/wgpu_helper/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod framework;
 pub mod transforms;
 pub mod vertex_data;
 pub mod wgpu_simplified;
