+++ practice/cube_blinn_phong/Cargo.toml
@@ -12,5 +12,5 @@ env_logger  = "0.11.5"
 wgpu        = "22.1.0"
 winit       = "0.30.5"
 wgpu_helper = { path = "../../lib/wgpu_helper" }
-pollster = "0.3.0"
-rand = "0.8.5"
+pollster    = "0.3.0"
+rand        = "0.8.5"
